# openapi.yaml
openapi: 3.1.0
info:
  title: FabzClean Modern Backend (Option B)
  version: "1.0.0"
  description: |
    OpenAPI 3.1 specification for the FabzClean modernized backend (Flask + MySQL).
    - JWT auth for customers
    - Session cookie for admin
    - Worker token auth (bearer)
servers:
  - url: https://{{host}}
    description: Production (replace {{host}} with your domain or static IP)
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token (Authorization: Bearer <token>). Used by customer endpoints."
    workerToken:
      type: http
      scheme: bearer
      bearerFormat: "worker-token"
      description: "Worker device token (Authorization: Bearer <worker_token>)."
    adminSession:
      type: apiKey
      in: cookie
      name: session
      description: "Admin session cookie (set after /admin/login)."
  schemas:
    Error:
      type: object
      properties:
        msg:
          type: string
      example:
        msg: "Bad credentials"
    Service:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
          format: float
        duration_minutes:
          type: integer
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, price]
    ServiceCreate:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
          format: float
        duration_minutes:
          type: integer
      required: [name, price]
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, email]
...
    Order:
      type: object
      properties:
        id:
          type: integer
        order_number:
          type: string
        customer_id:
          type: integer
        service_id:
          type: integer
        service_name:
          type: string
        pickup_date:
          type: string
          format: date-time
          nullable: true
        instructions:
          type: string
          nullable: true
        total_cost:
          type: number
          format: float
        status:
          type: string
          enum: [created, picked_up, processing, completed, delivered, cancelled]
        qr_code_path:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, order_number, customer_id, service_id, total_cost, status]
    OrderCreate:
      type: object
      properties:
        service_id:
          type: integer
        pickup_date:
          type: string
          format: date-time
        instructions:
          type: string
      required: [service_id]
    Track:
      type: object
      properties:
        id:
          type: integer
        order_id:
          type: integer
        worker_id:
          type: integer
          nullable: true
        action:
          type: string
        note:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [id, order_id, action, created_at]
...
paths:
...
  /api/v1/customers/:
    get:
      summary: Get current customer's profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Customer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update current customer's profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                phone:
                  type: string
              example:
                name: "New Name"
                phone: "9999999999"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/services/:
    get:
      summary: List services
      responses:
        "200":
          description: Array of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      summary: Create a service (admin or privileged)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceCreate"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "400":
          description: Validation/error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/services/{service_id}:
    parameters:
      - name: service_id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Get service by id
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
    put:
      summary: Update a service
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceCreate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      summary: Delete a service
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                example:
                  msg: "Deleted"
  /api/v1/orders/:
    get:
      summary: List orders for current customer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      summary: Create a new order (customer JWT required)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreate"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/orders/{order_id}:
    parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Get order by id (owner only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "403":
          description: Forbidden
        "404":
          description: Not found
    put:
      summary: Update an order (owner only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pickup_date:
                  type: string
                  format: date-time
                instructions:
                  type: string
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                example:
                  msg: "Updated"
        "403":
          description: Forbidden
    delete:
      summary: Delete an order (owner only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                example:
                  msg: "Deleted"
        "403":
          description: Forbidden
...
  /api/v1/tracks/order/{order_id}:
    get:
      summary: Get tracks for a specific order
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Array of tracks for order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
        "404":
          description: Order not found
...
  /admin/api/orders:
    get:
      summary: List all orders (admin session required)
      security:
        - adminSession: []
      responses:
        "200":
          description: Array of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Admin login required
security:
  - {}
...